+++ R/differential_expression.R
@@ -22,26 +22,17 @@
 #' @importFrom future.apply future_lapply
 #' @importFrom stats model.matrix p.adjust pchisq
 #'
-#' @examples
-#' \dontrun{
-#' vst_out <- vst(pbmc, return_cell_attr = TRUE)
-#' # create fake clusters
-#' clustering <- 1:ncol(pbmc) %/% 100
-#' res <- compare_expression(x = vst_out, umi = pbmc, group = clustering, val1 = 0, val2 = 3)
-#' }
-#'
 compare_expression <- function(x, umi, group, val1, val2, method = 'LRT', bin_size = 256,
                                cell_attr = x$cell_attr, y = x$y, min_cells = 5,
                                weighted = TRUE, randomize = FALSE, verbosity = 2,
-                               verbose = TRUE, show_progress = TRUE) {
+                               verbose = NULL, show_progress = NULL) {
   # Take care of deprecated arguments
-  args_passed <- names(sapply(match.call(), deparse))[-1]
-  if ('verbose' %in% args_passed) {
-    warning("The 'verbose' argument is deprecated as of v0.3. Use 'verbosity' instead.", immediate. = TRUE)
+  if (!is.null(verbose)) {
+    warning("The 'verbose' argument is deprecated as of v0.3. Use 'verbosity' instead. (in sctransform::vst)", immediate. = TRUE, call. = FALSE)
     verbosity <- as.numeric(verbose)
   }
-  if ('show_progress' %in% args_passed) {
-    warning("The 'show_progress' argument is deprecated as of v0.3. Use 'verbosity' instead.", immediate. = TRUE)
+  if (!is.null(show_progress)) {
+    warning("The 'show_progress' argument is deprecated as of v0.3. Use 'verbosity' instead. (in sctransform::vst)", immediate. = TRUE, call. = FALSE)
     if (show_progress) {
       verbosity <- 2
     } else {